+++ infra/prod/modules/app-registry/main.tf
@@ -1,0 +1,4 @@
+resource "aws_servicecatalogappregistry_application" "this" {
+  name        = var.app_name
+  description = "SC Hauling Planner PWA"
+}