+++ infra/prod/modules/app-registry/variables.tf
@@ -1,0 +1,5 @@
+variable "app_name" {
+  description = "AppRegistry application name"
+  type        = string
+  default     = "sc-haul"
+}