+++ infra/prod/modules/site/variables.tf
@@ -1,0 +1,5 @@
+variable "aws_account_id" {
+  description = "AWS account ID (used in bucket name)"
+  type        = string
+}
+